% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpm-package.R
\docType{package}
\name{mcmp}
\alias{mcpm}
\alias{mcpm-package}
\alias{mcmp-package}
\title{\code{mcmp} package}
\description{
Wrapper of an algorithm for Maximum Cardinality Matching and Minimum Cost Perfect Matching
Problems in General Graphs.
}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
