% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minimum_cost_perfect_matching}
\alias{minimum_cost_perfect_matching}
\title{Compute minimum cost perfect matching}
\usage{
minimum_cost_perfect_matching(nvertices, nedges, edges)
}
\arguments{
\item{nvertices}{The number of vertices in the graph.}

\item{nedges}{The number of edges in the graph.}

\item{edges}{A dataframe representing the edges and their cost with columns \code{from}, \code{to} and
\code{edges}.}
}
\value{
A list with entries \code{cost} (a number) and \code{edges} (a matrix).
}
\description{
Compute minimum cost perfect matching
}
